requirements.txt
Django>=4.2.13,<5
djangorestframework>=3.14,<4
djangorestframework-simplejwt>=5.5,<6
psycopg2-binary>=2.9,<3
dj-database-url>=2.1,<3
python-dotenv>=1.0,<2
